Cross-compile under Linux ready
This is used for development needs. Cross compile toolchain is mingw-w64.
Added development option to enable development test dll coexistence with the
original dll in N++ plugins dir.
Add GTags binaries dir variable that is separate from the plugin's name
as it was until now so it can be used by both dlls.