Add "Ignore Case" option command to the menu
"Match Case" from the Search window is now changed to "Ignore Case".
It overwrites the default "Ignore Case" option from the menu but is not saved.
The "Ignore Case" from the menu is used for the default searches (triggered without using the Search window).
Results window tab header info doesn't "literal" and "match case" anymore. It shows only "regexp" and/or "ignore case".