Force ignore case on Scintilla autocomplete
This is temporary for testing until a better solution can be added to
Settings window.  The settings dialog should provide:
 + [] enable Scintilla autocomplete
 + [] ignore case
 + [__] from the nth character

The character text box should accept 1-9 as valid characters.  Or it
could use a slider like the Notepad++ one, though that may be harder
to implement.

I think autocomplete ignore case should be different than definition
and reference ignore case.  With Def and Ref, I usually want the
exact case match as I see an object in the code, deliberately
click on it and then find the Def or Refs of that "exact" (i.e.,
case-sensitive) object.  With autocomplete, I'm getting assitance
from the "library" of available objects to help me find what I'm
looking for.  For example, is the "number of characters":
 + .length
 + .len
 + .Length
 + .Len
 + .size
 + is it a property, or does it end in `()` as a method, etc..

Autocomplete should provide more of a "fuzzy-find" (i.e., case-
insensitive) match.  Though others may think otherwise so there
should be an option to make is case-sensitive, but that option
should be different for "autocomplete" vs. "Defs and Refs".